fix: resolve bugs.md audit
- clippy op_ref/map_or/redundant_closure in registry/result_view/session_view/host/server
- harden decode_snmp_response bounds and add truncated input tests
- guard LDAP pos+1 reads at msg len and DN attr boundaries
- clamp brute PIN max_len to 9 to avoid u32 pow overflow
- remove panicking Default for AsyncNseExecutor
- add tracing error before panic on runtime_bridge fallback
- add PortScanResults.results_truncated and wire through ctors
- document is_private_ip as legacy broad
- widen trend severity_change to i64
- poison-resilient RwLock in eggsec-python
- wrap UDP fingerprint tasks with overall timeout
- avoid clone in mutator hot loop
- switch policy_decision to FxHashSet
- reuse reqwest Client via LazyLock in pipeline executor

All clippy -p eggsec/eggsec-nse/eggsec-ui-model/eggsec-daemon clean; 4152 tests pass; make check ALL PASSED

+++ crates/eggsec-output/src/trend.rs
@@ -96,9 +96,9 @@ impl ResultComparator {
         }
 
         let severity_change = (
-            new.summary.critical as i32 - old.summary.critical as i32,
-            new.summary.high as i32 - old.summary.high as i32,
-            new.summary.medium as i32 - old.summary.medium as i32,
+            new.summary.critical as i64 - old.summary.critical as i64,
+            new.summary.high as i64 - old.summary.high as i64,
+            new.summary.medium as i64 - old.summary.medium as i64,
         );
 
         ComparisonResult {
@@ -132,7 +132,7 @@ pub struct ComparisonResult {
     pub added: Vec<Finding>,
     pub removed: Vec<Finding>,
     pub unchanged: Vec<Finding>,
-    pub severity_change: (i32, i32, i32),
+    pub severity_change: (i64, i64, i64),
     pub old_timestamp: String,
     pub new_timestamp: String,
 }
